# app/openapi.yaml
openapi: 3.0.0
info:
  title: 'OpenApi Documentation For Portal Vagas'
  description: 'Documentation intended to guide the use of the API to access Portal Vagas data'
  contact:
    email: portal_vagas_admin@gmail.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8190'
    description: 'Demo API Server'
paths:
...
  /api/opportunities:
    get:
      tags:
        - Opportunities
      summary: 'Get list of opportunities'
      description: 'Returns list of opportunities'
      operationId: getOpportunitiesList
      parameters:
        -
          $ref: '#/components/parameters/page'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResourceCollection'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    post:
      tags:
        - Opportunities
      summary: 'Store new opportunity'
      description: 'Returns opportunity data'
      operationId: storeOpportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportunityRequest'
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
  '/api/opportunities/{id}':
    get:
      tags:
        - Opportunities
      summary: 'Get opportunity information'
      description: 'Returns opportunity data'
      operationId: getOpportunityById
      parameters:
        -
          name: id
          in: path
          description: 'Opportunity id'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    put:
      tags:
        - Opportunities
      summary: 'Update existing opportunity'
      description: 'Returns updated opportunity data'
      operationId: updateOpportunity
      parameters:
        -
          name: id
          in: path
          description: 'Opportunity id'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOpportunityRequest'
      responses:
        '202':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Resource Not Found'
    delete:
      tags:
        - Opportunities
      summary: 'Delete existing opportunity'
      description: 'Deletes a record and returns no content'
      operationId: deleteOpportunity
      parameters:
        -
          name: id
          in: path
          description: 'Opportunity id'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema: {  }
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Resource Not Found'
components:
  schemas:
...
    CreateOpportunityRequest:
      properties:
        contract_type:
          type: string
          example: CLT
        accepts_applications:
          type: boolean
          example: 'false'
        offered_salary:
          type: integer
          example: '18237'
      type: object
...
    UpdateOpportunityRequest:
      properties:
        contract_type:
          type: string
          example: CLT
        accepts_applications:
          type: boolean
          example: 'false'
        offered_salary:
          type: integer
          example: '18237'
      type: object
...
    OpportunityResource:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Sample Name'
      type: object
    OpportunityResourceCollection:
      type: array
      items:
        $ref: '#components/schemas/ApplicationResource'
  parameters:
    page:
      name: page
      in: query
      description: 'A pagina alvo'
      required: true
      schema:
        type: integer
        example: '1'
...